requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.17.0
tensorflow-hub==0.16.1
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9